[fixed] TC fixes
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@840 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/login/LoginPacketHandler.cpp
@@ -162,7 +162,7 @@ void LoginPacketHandler::handleDeleteCharacterMessage(Message* pack) {
 
 	uint32 ServerId = pack->parseInt();
 	pack->shiftOffset(4);
-    uint64 charId = pack->parseInt();
+    uint32 charId = pack->parseInt();
     int dbDelete;
     string firstName;
